Struct sql_ops::partition::PartitionedRelation

pub struct PartitionedRelation<T: DeviceCopy> {
    pub relation: Mem<T>,
    pub offsets: Mem<u64>,
    /* private fields */
}

A radix-partitioned relation, optionally with padding in front of each partition.

The relation supports chunking on a single GPU. E.g. in the Chunked algorithm, there is a chunk per thread block. In this case, chunks should equal the grid size.

Invariants

len must match in PartitionOffsets

radix_bits must match in GpuRadixPartitioner.

max_chunks must equal the maximum number of chunks computed at runtime (e.g., the grid size).

Fields
relation: Mem<T>offsets: Mem<u64>
Implementations

impl<T: DeviceCopy> PartitionedRelation<T>

pub fn new(
len: usize,
histogram_algorithm_type: HistogramAlgorithmType,
radix_bits: u32,
max_chunks: u32,
partition_alloc_fn: MemAllocFn<T>,
offsets_alloc_fn: MemAllocFn<u64>
) -> Self

Creates a new partitioned relation, and automatically includes the necessary padding and metadata.

pub fn len(&self) -> usize

Returns the total number of elements in the relation (excluding padding).

pub fn padded_len(&self) -> usize

Returns the total number of elements in the relation including padding.

pub fn bytes(&self) -> usize

Returns the total number of bytes used to store the relation (including metadata).

pub fn capacity(&self) -> usize

Returns the number of elements allocated in memory (excluding padding).

The capacity includes unused elements, but excludes padding.

pub fn resize(&mut self, new_len: usize) -> Result<()>

Resizes the PartitionedRelation in-place so that len is equal to new_len.

If new_len is greater than the allocated memory capacity, then the resize will abort and return Err.

Post-conditions

The resize invalidates all data contained in relation and the offsets.

However, an aborted resize leaves PartitionedRelation intact and unmodified.

pub fn num_chunks(&self) -> u32

Returs the number of chunks.

pub fn fanout(&self) -> u32

Returns the number of partitions.

pub fn radix_bits(&self) -> u32

Returns the number of radix bits.

pub fn partition_len(&self, partition_id: u32) -> Result<usize>

Returns the length of the requested partition.

If the offsets are accessible by the CPU (i.e., in DerefMem), then the length is returned. Otherwise, the function returns an error.

pub fn padding_len(&self) -> u32

Returns the number of padding elements per partition.

pub unsafe fn as_raw_relation_slice(&self) -> Result<&[T]>

Returns the internal representation of the relation data as a slice.

This function is intended for unit testing. Use the methods provided by the Index trait or chunks_mut() instead if possible.

The function is unsafe because:

the internal representation may change

padding may contain uninitialized memory.

pub unsafe fn as_raw_relation_mut_slice(&mut self) -> Result<&mut [T]>

Returns the internal representation of the relation data as a mutable slice.

This function is intended for unit testing. Use the methods provided by the Index trait or chunks_mut() instead if possible.

The function is unsafe because:

the internal representation may change

padding may contain uninitialized memory.

impl<K: DeviceCopy, V: DeviceCopy> PartitionedRelation<Tuple<K, V>>

pub fn chunks_mut(&mut self) -> PartitionedRelationChunksMut<'_, Tuple<K, V>>

Returns an iterator over the chunks contained inside the relation.

Chunks are non-overlapping and can safely be used for parallel processing.